Title: n개의 숫자 중 최소, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240809/n개의 숫자 중 최소/min-of-n-num.cpp
@@ -21,7 +21,7 @@ int main() {
 	for (int i = 0; i < n; i++)
 	{
 		int num; cin >> num;
-		mi = min(mi, num);
+		mi = min(mi,(ll)num);
 	}
 	cout << mi;
 	for (int i = 0; i < n; i++)
